+++ src/components/index.ts
@@ -16,4 +16,4 @@ export { Loading } from './Loading'
 
 export { MapView } from './MapView'
 
-export { MapViewLeaf } from './MapViewLeaf'
+export { MapViewLeaf } from './MapLeaflet/MapViewLeaf'
