+++ src/context/index.ts
@@ -3,3 +3,6 @@ export { MapProvider } from './map/MapProvider'
 
 export { PlacesContext } from './places/PlacesContext'
 export { PlacesProvider } from './places/PlacesProvider'
+
+export { ModalContext } from './modal/ModalContext'
+export { ModalProvider } from './modal/ModalProvider'
